IMPALA-11356: Fix test_sfs.py to account for new Hive behavior
Hive changed behavior in HIVE-24920 to allow external SFS tables to
point to locations in managed tables. test_sfs.py had a test case for
the previous failure, so it broke with this change.

This changes test_sfs.py to verify the new behavior where an external
SFS table can point to a managed location.

Testing:
 - Ran test_sfs.py locally with recent GBN

Change-Id: I426aacdba7afba3f3a747a4e02632d15e38c63c9
Reviewed-on: http://gerrit.cloudera.org:8080/18618
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>