IMPALA-11106: Make Impala compatible with Iceberg 0.13
Currently we are using deprecated methods from Iceberg API which won't
be in 0.13. We need to switch to the newer methods:

* PartitionSpecVisitor.visit
* HiveCatalog constructor

We are also using SnapshotUtil from the Iceberg Core package. One of its
methods has changed semantically. We shouldn't depend on packages
outside of Iceberg API as they are treated as internal modules.

Testing:
* built Impala with Iceberg 0.13.0 and executed Iceberg-related tests

Change-Id: I7a95483ea3ce15af9a344f9da3b7bfb7dea8a10a
Reviewed-on: http://gerrit.cloudera.org:8080/18196
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>