IMPALA-11109: Catch class loading error for UDFs.
Fixes regression caused by IMPALA-10997.

If a UDF fails to load, an exception needs to be caught
and logged or else catalogd will not start up. The
method HiveLegacyJavaFunction.extract() will catch any hidden
exception thrown, and rethrow it as a CatalogException. The caller
already handles the CatalogException and logs the message
appropriately. The specific "hidden" exception thrown in our
regression tests was a "ClassNotFoundException", but in case
there are other exceptions that are thrown, we catch the generic
"Throwable" exception because the "extract" method should not
prevent the server from coming up.

Change-Id: I16813209cd4c2367c45e569c2aac13eff7ce2dbf
Reviewed-on: http://gerrit.cloudera.org:8080/18207
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>