IMPALA-10989: fix race for result set metadata
TSAN tests uncovered a race condition between the
thread reading the result set metadata in
ImpalaServer::GetResultSetMetadata() and the
thread setting the result set metadata in
ClientRequestState::SetResultSet() from
ClientRequestState::ExecDdlRequestImpl().
This is introduced by IMPALA-10811, which runs
ExecDdlRequestImpl in an async thread that
can now race with the client thread.

GetResultSetMetadata() holds ClientRequestState's
lock_ while reading the result set metadata, so
the fix is to hold this lock when writing the
result set metadata.

Testing:
 - Ran TSAN core job

Change-Id: Ic0833ed20d62474c434fa94bbbf8cd8ea99a7cf4
Reviewed-on: http://gerrit.cloudera.org:8080/18212
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>