IMPALA-11133: Decode author of a commit with utf8 before printing it
We found that compare_branches.py could fail if the author of a commit
contains non-ASCII characters because the script attempts to print the
field. This patch fixes the problem by explicitly decoding the value of
author with the encoding 'utf8'. The commit message is also decoded with
'utf8' to prevent similar problems from happening when there are
non-ASCII characters in the commit message.

Testing:
 - Manually verified that we won't get the UnicodeDecodeError after this
   patch.

Change-Id: Ieb03b0937a994db2bf08e4199574d04f7fb99f5d
Reviewed-on: http://gerrit.cloudera.org:8080/18256
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>