IMPALA-10999 Flakiness in TestAsyncLoadData.test_async_load
This patch addresses the flakness by allowing RUNNING state to be a
legit exec state returned from execute_query_async_using_client() in
python. This call submits the load query to Impala backend.

The corresponding Impala backend code for the beewax protocol is
ImpalaServer::query() which utilizes a wait thread executing
ClientRequestState::Wait() to set the exec state from RUNNING to
FINISH. Sometime, when this wait thread does not run fast to do so,
the returned state from the main thread is RUNNING.

The fix is purely a modification to the test itself.

Testing:
  1. ran core test successfully

Change-Id: Ic2ac954b0494b7413ce0ec405718fcc354dba9e0
Reviewed-on: http://gerrit.cloudera.org:8080/18268
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>