IMPALA-11106 (part 2): Add "uuid" to the ignored tbl properties in test_show_create_table.py
The first patch of IMPALA-11106 adds compatibility with Iceberg 0.13.
With Impala is able to compile with both Iceberg 0.12 and 0.13, and
the tests of test_iceberg.py execute successfully.

Though the test test_show_create_table.py still fails because Iceberg
0.13 adds a new table property to the table: "uuid". The value of it
cannot be determined so I added "uuid" to the ignored table properties.

Testing:
 * tested manually by using Iceberg 0.13

Change-Id: Ia6ad0cdfb6acdd1e3adcd2035b6104291e346ef0
Reviewed-on: http://gerrit.cloudera.org:8080/18273
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>