IMPALA-11156: TestHmsIntegration.test_desc_json_table failed in exhaustive build
The test was getting skipped as it was not in exhaustive strategy,
to fix that its been moved to TestHmsIntegrationSanity(ImpalaTestSuite)
class. Now there are two tests, test_desc_json_tbl and
test_desc_json_tbl_hive to check whether JSON tables created in
impala can be described in hive and vice versa.
Change-Id: I4ddd983aeb82817eb355ca94dcf4a6875986f8e1
Reviewed-on: http://gerrit.cloudera.org:8080/18299
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>