IMPALA-11176: Fix Memory leak in ClientCacheHelper
The patch fixes a memory leak issue in client cache.

The memory leak happens because a shared pointer of the
base class "ThriftClientImpl" is created from a derived class
"ThriftClient" while the base class's destructor is non-virtual.
As a result, the shared pointer can't locate the correct
destructor, which should be the derived one, and leads to
memory leaks.

The way to fix in the patch is to make the destructor of
"ThriftClientImpl" virtual. Added testcase ClientCacheTest.

Tests:
Ran exhaustive tests.

Change-Id: I2bf3d7e7d897c89eb4538df5382ce3c590055b71
Reviewed-on: http://gerrit.cloudera.org:8080/18314
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>