IMPALA-111218: TestIcebergTable.test_table_load_time_for_many_files flaky
In some test runs test_table_load_time_for_many_files failed because
the time limit (3 seconds) was too short. In the failed tests it was
around 4 seconds to execute the DESCRIBE statement.

This patch increases the time limit to 10 seconds when HDFS is used.
This should be long enough to fix the flakiness, but short enough
to catch pathologically long table loads (the one that was fixed by
IMPALA-11214).

Change-Id: Ic48bb561e97c008c6f4e51fc9d9ec8b3f7cdaf9e
Reviewed-on: http://gerrit.cloudera.org:8080/18376
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>